Partition and refactor
