+++ src/LinqExercises/Features/AdvancedLINQPatterns/IAdvancedLINQPattern.cs
@@ -1,0 +1,7 @@
+namespace LinqExercises.Features.AdvancedLINQPattern
+{
+    public interface IAdvancedLINQPatternService
+    {
+        Task RunAdvancedLINQPatternExerciseAsync();
+    }
+}
