+++ src/LinqExercises/Features/Partition/IPartitionService.cs
@@ -1,0 +1,7 @@
+namespace LinqExercises.Features.Partition
+{
+    public interface IPartitionService
+    {
+        Task RunPartitionExerciseAsync();
+    }
+}
