+++ src/LinqExercises/Features/Quantifier/QuantifierService.cs
@@ -17,16 +17,38 @@ public class QuantifierService : IQuantifierService
     {
         _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - Running quantifier exercises...");
         // Basic Any
+        _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - Basic Any Example");
         var hasLongTrackTitles = _dbContext.Tracks.Any(t => t.Name!.Length > 50);
-        _logger.LogInformation($"Are there any tracks with titles longer than 50 characters? {hasLongTrackTitles}");
+        var anyTable = new Spectre.Console.Table().Title("[green]Any Quantifier Result[/]").Border(Spectre.Console.TableBorder.Rounded)
+            .AddColumn("Has Track Titles Longer Than 50 Characters");
+        anyTable.AddRow(hasLongTrackTitles.ToString());
+        AnsiConsole.Write(anyTable);
+        AnsiConsole.MarkupLine("[bold italic]Press any key to continue to the next Quantifier example...[/]");
+        Console.ReadKey();
+        
         // Basic All
+        _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - Basic All Example");
         var allTracksHaveComposer = _dbContext.Tracks.All(t => !string.IsNullOrEmpty(t.Composer));
-        _logger.LogInformation($"Do all tracks have a composer? {allTracksHaveComposer}");
+        var allTable = new Spectre.Console.Table().Title("[green]All Quantifier Result[/]").Border(Spectre.Console.TableBorder.Rounded)
+            .AddColumn("Do All Tracks Have a Composer?");
+        allTable.AddRow(allTracksHaveComposer.ToString());
+        AnsiConsole.Write(allTable);
+        AnsiConsole.MarkupLine("[bold italic]Press any key to continue to the next Quantifier example...[/]");
+        Console.ReadKey();
+
         // Basic Contains
+        _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - Basic Contains Example");
         var trackNames = _dbContext.Tracks.Select(t => t.Name).ToList();
         var containsSpecificTrack = trackNames.Contains("Imagine");
-        _logger.LogInformation($"Does the track list contain 'Imagine'? {containsSpecificTrack}");
+        var containsTable = new Spectre.Console.Table().Title("[green]Contains Quantifier Result[/]").Border(Spectre.Console.TableBorder.Rounded)
+            .AddColumn("Does the Track List Contain 'Imagine'?");
+        containsTable.AddRow(containsSpecificTrack.ToString());
+        AnsiConsole.Write(containsTable);
+        AnsiConsole.MarkupLine("[bold italic]Press any key to continue to the next Quantifier example...[/]");
+        Console.ReadKey();
+        
         // Concat
+        _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - Concat Example");
         var playlist1TrackIds = _dbContext.PlaylistTracks
             .Where(pt => pt.PlaylistId == 1)
             .Select(pt => pt.TrackId);
@@ -34,11 +56,15 @@ public class QuantifierService : IQuantifierService
             .Where(pt => pt.PlaylistId == 2)
             .Select(pt => pt.TrackId);
         var combinedTrackIds = playlist1TrackIds.Concat(playlist2TrackIds).Distinct();
-        _logger.LogInformation($"Combined track IDs from playlists 1 and 2 contains {combinedTrackIds.Count()} unique tracks.");
-
-        _logger.LogInformation("Press any key to continue to the next quantifier example..."); 
+        var concatTable = new Spectre.Console.Table().Title("[green]Combined Track IDs from Playlists 1 and 2[/]").Border(Spectre.Console.TableBorder.Rounded)
+            .AddColumn("Track ID");
+        combinedTrackIds.ToList().ForEach(trackId => concatTable.AddRow(trackId.ToString()));        
+        Spectre.Console.AnsiConsole.Write(concatTable);        
+        AnsiConsole.MarkupLine("[bold italic]Press any key to continue to the next Quantifier example...[/]");
         Console.ReadKey();
+        
         // ToDictionary
+        _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - ToDictionary Example");
         var trackDictionary = _dbContext.Tracks
             .Where(t => combinedTrackIds.Contains(t.TrackId))
             .ToDictionary(t => t.TrackId, t => t.Name);
@@ -52,7 +78,7 @@ public class QuantifierService : IQuantifierService
             trackDictionaryTable.AddRow(key.ToString(), Spectre.Console.Markup.Escape(trackDictionary[key] ?? string.Empty));
         }
         Spectre.Console.AnsiConsole.Write(trackDictionaryTable);
-        _logger.LogInformation("Press any key to finish the quantifier exercises...");
+        AnsiConsole.MarkupLine("[bold italic]Press Enter to continue to finish Quantifier Exercise...[/]");
         Console.ReadKey();
         _logger.LogInformation("[QuantifierService.RunQuantifierExerciseAsync] - Completed quantifier exercises.");
     }
